DROTH-2706 change to numeric so desimal is possible

+++ digiroad2-oracle/src/main/resources/db/migration/V0_1__create_tables.sql
@@ -181,7 +181,7 @@ CREATE TABLE enumerated_value
 (
     id            bigint       NOT NULL,
     property_id   bigint       NOT NULL,
-    value         bigint,
+    value         numeric,
     name_fi       varchar(512) NOT NULL,
     name_sv       varchar(512),
     image_id      bigint,
